// openmetadata-ui/src/main/resources/ui/src/components/DataQuality/DataQualityIndicator/DataQualityIndicator.types.ts
/*
 *  Copyright 2026 Collate.
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *  http://www.apache.org/licenses/LICENSE-2.0
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 */
export interface DataQualityIndicatorCounts {
  failingTests: number;
  unresolvedIncidents: number;
  upstreamIssues: number;
}

export type DataQualityIndicatorLevel =
  | 'failing'
  | 'incident'
  | 'upstream'
  | 'none';

export interface DataQualityIndicatorProps {
  counts: DataQualityIndicatorCounts;
  tableFqn: string;
}